refactor: Часть кода из компонента reviews вынесена в отдельный компонент
